calculate prev and next exercises
